postcode search combine model
