+++ src/Defra.Trade.ReMoS.AssuranceService.UI.Hosting/Pages/Shared/Establishments/PostcodeSearchPartial.cshtml
@@ -1,0 +1,19 @@
+﻿@using Defra.Trade.ReMoS.AssuranceService.UI.Hosting.Constants;
+
+
+<legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
+    <h1 class="govuk-fieldset__heading">
+        <span class="govuk-caption-l">Establishments</span>
+        @Model.ContentHeading
+    </h1>
+</legend>
+
+<p class="govuk-body govuk-hint">@Model.ContentText</p>
+
+<p class="govuk-body govuk-hint">You need to provide the details of all the establishments your business uses so they can be added when applying for General Certificates.</p>
+
+<p class="govuk-body govuk-inset-text">
+    All businesses moving, handling or receiving consignments under the scheme must comply with the <a asp-page="@Routes.Pages.Path.TermsAndConditionsPath" rel="noopener noreferrer" target="_blank">terms and conditions (opens in a new tab)</a> of the scheme.
+    Competent authorities may conduct checks on any consignments.
+</p>
+
